bot: hugely simplified parser
Because entry text is now quoted, we can use it as a single lexical
entity, therefore getting rid of cubersome "word" lexeme definition,
which tends to clash in many ways with other lexemes and their parsing
results. This way entry text remains untangled by the parser and being
saved to DB "as is". Doing so allows us to use almost any special symbol
we need within entry contents as far as the quoting is correct. On the
other hand it allows to keep "command language" and user data separated
in terms of parsing definitions.